# [TESTS] TSDB: Faster WAL benchmarks