* Reuse buffer for writing the WAL records. Less garbage collection, goes faster.
* More efficient label creation - use the Builder abstraction instead of going via a map.

Before:
```
go test -run XXX -bench 'BenchmarkLoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=
0.300,oooCapMax=32,missingSeriesPct=0.000' ./tsdb/
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Core(TM) i7-14700K
BenchmarkLoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32,missingSeriesPct=0.000-28                    1        3104853168 ns/op
PASS
ok      github.com/prometheus/prometheus/tsdb   4.310s
```

After:
```
go test -run XXX -bench 'BenchmarkLoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32,missingSeriesPct=0.000' ./tsdb/
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Core(TM) i7-14700K
BenchmarkLoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32,missingSeriesPct=0.000-28                    1        3024799543 ns/op
PASS
ok      github.com/prometheus/prometheus/tsdb   3.863s
```

It makes even more of a difference if you push the number of series up.